Remove special variable support
Variables containing character such as +*%#!~ should not be considered as legal values. This is still possible to assign them in a map using get and set functions, but supporting them directly into the Razor syntax is just error prone.

Assignment such as `@{a.b} += 1` must be handled like global variables. We should add (according to StrictAssignationMode) the same kind of validation since it assigns value in a map.